requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
tortoise-orm==0.19.3
aerich==0.7.2
psycopg2-binary==2.9.9
python-multipart==0.0.6
email-validator==2.0.0
pydantic==1.10.12
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0 